refactor: customRoutes와 reactAdmin 코드 분리를 가시적으로 하기 위해 디렉터리 정리
