--- src/components/custom/Roadmap/index.tsx
@@ -1,11 +1,0 @@
-import RoadmapSessions from './RoadmapSessions';
-
-const Roadmap = () => {
-  // 세션들을 가져와서 보여주고 선택할 수 있게 한다.
-
-  // 선택된 세션id로부터 1뎁스 키워드 리스트를 가져와서 보여준다
-
-  return <RoadmapSessions />;
-};
-
-export default Roadmap;
